[TRL-283] refactor: TripQueryRepository 구조 개선
