+++ src/main/java/com/cosain/trilo/trip/infra/repository/trip/TripQueryRepository.java
@@ -1,16 +1,66 @@
 package com.cosain.trilo.trip.infra.repository.trip;
 
+import com.cosain.trilo.trip.infra.dto.QTripDetail;
+import com.cosain.trilo.trip.infra.dto.QTripSummary;
 import com.cosain.trilo.trip.infra.dto.TripDetail;
 import com.cosain.trilo.trip.infra.dto.TripSummary;
+import com.querydsl.jpa.impl.JPAQuery;
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
+import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-public interface TripQueryRepository {
-    Optional<TripDetail> findTripDetailByTripId(Long tripId);
+import static com.cosain.trilo.trip.domain.entity.QTrip.trip;
 
-    Slice<TripSummary> findTripSummariesByTripperId(Long tripperId, Pageable pageable);
+@Repository
+@RequiredArgsConstructor
+public class TripQueryRepository{
 
-    boolean existById(Long tripI);
+    private final JPAQueryFactory query;
+
+    /**
+     * Entity 가 아닌 DTO 로 필요한 속성들만 조회, 프로젝션 활용
+     * JPQL 을 사용할 수도 있으나,
+     * 1. new 키워드 사용
+     * 2. package 이름 다 적어야돼서 지저분함
+     * 3. 생성자만 지원
+     * 등의 단점을 극복하기 위해 QueryDSL 사용
+     *
+     */
+    public Optional<TripDetail> findTripDetailById(Long tripId){
+
+        return Optional.ofNullable(query.select(
+                        new QTripDetail(trip.id, trip.tripperId, trip.tripTitle.value, trip.status, trip.tripPeriod.startDate, trip.tripPeriod.endDate))
+                .from(trip)
+                .where(trip.id.eq(tripId))
+                .fetchOne());
+    }
+
+    public Slice<TripSummary> findTripSummariesByTripperId(Long tripperId, Pageable pageable) {
+        JPAQuery<TripSummary> jpaQuery = query.select(new QTripSummary(trip.id, trip.tripperId, trip.tripTitle.value, trip.status, trip.tripPeriod.startDate, trip.tripPeriod.endDate))
+                .from(trip)
+                .where(trip.tripperId.eq(tripperId))
+                .orderBy(trip.id.desc())
+                .offset(pageable.getOffset())
+                .limit(pageable.getPageSize());
+
+        long totalCount = query.from(trip)
+                .where(trip.tripperId.eq(tripperId))
+                .fetchCount();
+
+        return new PageImpl<>(jpaQuery.fetch(), pageable, totalCount);
+    }
+
+    public boolean existById(Long tripId) {
+        Integer fetchOne = query.selectOne()
+                .from(trip)
+                .where(trip.id.eq(tripId))
+                .fetchFirst();
+
+        return fetchOne != null;
+    }
 }
