--- src/main/java/com/cosain/trilo/trip/infra/repository/trip/TripQueryRepositoryImpl.java
@@ -1,33 +1,0 @@
-package com.cosain.trilo.trip.infra.repository.trip;
-
-import com.cosain.trilo.trip.infra.dto.TripDetail;
-import com.cosain.trilo.trip.infra.dto.TripSummary;
-import com.cosain.trilo.trip.infra.repository.trip.jpa.TripQueryJpaRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-@RequiredArgsConstructor
-public class TripQueryRepositoryImpl implements TripQueryRepository {
-
-    private final TripQueryJpaRepository tripQueryJpaRepository;
-
-    @Override
-    public Optional<TripDetail> findTripDetailByTripId(Long tripId) {
-        return tripQueryJpaRepository.findTripDetailById(tripId);
-    }
-
-    @Override
-    public Slice<TripSummary> findTripSummariesByTripperId(Long tripperId, Pageable pageable) {
-        return tripQueryJpaRepository.findTripSummariesByTripperId(tripperId, pageable);
-    }
-
-    @Override
-    public boolean existById(Long tripId) {
-        return tripQueryJpaRepository.existsById(tripId);
-    }
-}
